email_analyzer/requirements.txt
azure-identity
microsoft-graph-core
msgraph-sdk
